Update dependencies
Related Issues: #148

Updated the dependencies in `test/environment.yml` as the minimum
python version has been bumped to 3.9 and `rdkit` is not being
installed through `pypi`. `mdanalysis` is also being installed through
`conda install -c conda-forga mdanalysis` rather than
`pip install mdanalysis` because the `pip` version seem lacking
dependencies and is causing crashing when `pytest` is run. This is
fixed by simply using the `conda` version of the package, presumably
because it installs the required binaries that are missing in the
`pip` version.

Also updated `readme.rst` to reflect that `mdanalysis` needs to be
installed from `conda` and that `stk` and `rdkit` do not need to be
installed explicitly, as they are requirements in `setup.py`.

Finally, removed the `prefix` and `name` keys from
`tests/environment.yml` as these can modify global environments on the
users computer. Instead the environmnent file should be used in an
active environment in the following way

```
conda env update --file tests/environment.yml
```

which will simply add the dependencies in `tests/environment.yml` to
the current active environment, instead of creating a new global
enivonrment, as was being done previously.

Co-authored-by: Austin Mroz <austin.mroz@gmail.com>